+++ CHANGELOG.md
@@ -2,10 +2,32 @@
 
 ## 22.0-SNAPSHOT
 
+### 💥 Breaking Changes (upgrade difficulty: 🟢 LOW)
+* All `Timer`, `Cache`, and `CachedValue` object require a 'name' property.  This property was
+previously optional in many cases, but is now required in order to support new cluster features,
+logging, and admin tools.  The new `BaseService.resources` property now will give access to all
+resources by name, if needed and replaces `BaseService.timers`.
+
+* `BaseService` methods `getIMap()`, `getReplicatedMap()` and `getISet()` have been changed to
+  `createIMap()`, `createReplicatedMap()` and `createISet()`, respectively.  This change provides
+  a consistent interface for all resources on BaseService and is not expected to impact most
+ applications.
+
+### 🎁 New Features
+* `Cache` and `CachedValue` should now be created using a factory on `BaseService`.  This streamlined
+interface reduces boilerplate, and provides a consistent interface with `Timer`.
+
 ### ⚙️ Technical
+
+* Improvements to `Timer` to avoid extra executions when primary instance changes.
+
+* Updated `ClusterService` to use Hoist's `InstanceNotFoundException` class to designate routine.
 
 * Exposed `/xh/ping` as whitelisted route for basic uptime/reachability checks. Retained legacy
   `/ping` alias, but prefer this new path going forward.
+
+* Improvements to `RestController` to better support editing Domain Objects defined with secondary
+  domain objects.
 
 ## 21.0.1 - 2024-09-05